Refactored conversions of MatchConfig. Allow read/write of MatchConfig to JSON.
